\name{KinaseDownstream-package}
\alias{KinaseDownstream-package}
\alias{KinaseDownstream}
\docType{package}
\title{
KinaseDownstream: Visualization of Kinase Enrichment and Downstream Networks
}
\description{
KinaseDownstream provides functions to visualize significantly enriched kinases in phosphoproteomics data, the networks between each kinase, their downstream targets, and their immediate STRINGdb neighbors observed in matching proteomics data. This facilitates understanding of kinase interactions and their biological implications.
}
\details{
KinaseDownstream offers tools for visualizing kinase enrichment and downstream networks from phosphoproteomics and proteomics data.
}
The most important functions are:
\itemize{
  \item \code{\link{KinaseNetwork4substrates}}: Visualize kinase-substrate networks.
  \item \code{\link{processPTMSEAresult}}: Process PTMSEA results and generate summary plots.
  \item \code{\link{ppiNetwork4substrates}}: Build and visualize protein-protein interaction networks for kinase substrates.
}

\author{
Yuan Li, NBIS, SciLifeLab, Sweden
}

Maintainer: Yuan Li <yuan.li@nbis.se>

\references{

}

